components/requirements.txt
# ASR & Language ID requirements for Voxi
openai-whisper
torch
pydub